+++ packages/terra-breakpoints/src/terra-dev-site/test/breakpoints/ActiveBreakpointProviderTest.test.jsx
@@ -6,7 +6,17 @@ const ActiveBreakpointConsumer = withActiveBreakpoint(({ activeBreakpoint }) => 
     <p>
       The active breakpoint is:
       {' '}
-      {activeBreakpoint}
+      {activeBreakpoint.activeBreakpoint}
+      <br></br>
+      <br></br>
+      The event ('resize'/'orientationchange'):
+      {' '}
+      {activeBreakpoint.source}
+      <br></br>
+      <br></br>
+      The innerWidth:
+      {' '}
+      {activeBreakpoint.innerWidth}
     </p>
   </div>
 ));